Major refactor   - Magic Collections is now Magic List   - Removed `dict` and `str` completely
