+++ .gitignore
@@ -128,8 +128,5 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
-# Plates Bytecode files
-*.pbc
-
 # Custom
 \#*
