--- magic_collections/__init__.py
@@ -1,1 +1,0 @@
-from magic_collections.prelude import *  # noqa: F403
