+++ magic_list/__init__.py
@@ -1,0 +1,1 @@
+from magic_list.prelude import *  # noqa: F403
