--- tests/str_test.py
@@ -1,4 +1,0 @@
-# type: ignore
-import pytest
-
-from magic_collections import str
